refactor(vm): 统一导入指令类型并重命名GC方法
- 将 `bytecode::decoder::Instruction` 的导入路径统一更新为 `bytecode::instruction::Instruction`
- 将GC相关方法 `collect` 重命名为 `full_gc` 以更清晰地表达其行为
- 修复闭包创建函数调用中类型不匹配的问题

+++ projects/nyar-vm/src/vm/ops/closure.rs
@@ -23,7 +23,7 @@ impl NyarVM {
                     };
                     captured.push(Upvalue(val));
                 }
-                let v = Value::closure(module_idx, idx as usize, captured, &self.gc);
+                let v = Value::closure(module_idx, idx, captured, &self.gc);
                 self.push(v);
             }
             Instruction::LoadUpvalue(idx) => {
